+++ docs/.vitepress/config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "vitepress";
-import { en } from "./en";
-import { zh } from "./zh";
-import { shared } from "./shared";
+import en from "./en";
+import zh from "./zh";
+import shared from "./shared";
 
 // https://vitepress.dev/reference/site-config
 export default defineConfig({